# groot/rtree{,/rfunc}: introduce rfunc.Formula protocol