--- a/groot/rtree/example_reader_test.go
+++ b/groot/rtree/example_reader_test.go
@@ -10,6 +10,7 @@ import (
 
 	"go-hep.org/x/hep/groot"
 	"go-hep.org/x/hep/groot/rtree"
+	"go-hep.org/x/hep/groot/rtree/rfunc"
 )
 
 func ExampleReader() {
@@ -265,7 +266,7 @@ func ExampleReader_withFormulaFunc() {
 	err = r.Read(func(ctx rtree.RCtx) error {
 		v64 := f1()
 		str := f2()
-		fmt.Printf("evt[%d]: %v, %v, %v -> %g %g | %s\n", ctx.Entry, data.V1, data.V2, data.V3, f64.Eval(), v64, str)
+		fmt.Printf("evt[%d]: %v, %v, %v -> %g | %s\n", ctx.Entry, data.V1, data.V2, data.V3, v64, str)
 		return nil
 	})
 	if err != nil {
@@ -273,8 +274,143 @@ func ExampleReader_withFormulaFunc() {
 	}
 
 	// Output:
-	// evt[0]: 1, 1.1, uno -> 1311 1311 | "one": 1, "two": 1.1, "three": uno
-	// evt[1]: 2, 2.2, dos -> 2322 2322 | "one": 2, "two": 2.2, "three": dos
-	// evt[2]: 3, 3.3, tres -> 3433 3433 | "one": 3, "two": 3.3, "three": tres
-	// evt[3]: 4, 4.4, quatro -> 4644 4644 | "one": 4, "two": 4.4, "three": quatro
+	// evt[0]: 1, 1.1, uno -> 1311 | "one": 1, "two": 1.1, "three": uno
+	// evt[1]: 2, 2.2, dos -> 2322 | "one": 2, "two": 2.2, "three": dos
+	// evt[2]: 3, 3.3, tres -> 3433 | "one": 3, "two": 3.3, "three": tres
+	// evt[3]: 4, 4.4, quatro -> 4644 | "one": 4, "two": 4.4, "three": quatro
+}
+
+type UsrF64 struct {
+	rvars []string
+	v1    *int32
+	v2    *float32
+	v3    *string
+	fct   func(int32, float32, string) float64
+}
+
+var (
+	_ rfunc.Formula = (*UsrF64)(nil)
+)
+
+func (usr *UsrF64) RVars() []string { return usr.rvars }
+func (usr *UsrF64) Bind(args []interface{}) error {
+	if got, want := len(args), 3; got != want {
+		return fmt.Errorf(
+			"rfunc: invalid number of bind arguments (got=%d, want=%d)",
+			got, want,
+		)
+	}
+	usr.v1 = args[0].(*int32)
+	usr.v2 = args[1].(*float32)
+	usr.v3 = args[2].(*string)
+	return nil
+}
+
+func (usr *UsrF64) Func() interface{} {
+	return func() float64 {
+		return usr.fct(*usr.v1, *usr.v2, *usr.v3)
+	}
+}
+
+type UsrStr struct {
+	rvars []string
+	v1    *int32
+	v2    *float32
+	v3    *string
+	fct   func(int32, float32, string) string
+}
+
+var (
+	_ rfunc.Formula = (*UsrStr)(nil)
+)
+
+func (usr *UsrStr) RVars() []string { return usr.rvars }
+func (usr *UsrStr) Bind(args []interface{}) error {
+	if got, want := len(args), 3; got != want {
+		return fmt.Errorf(
+			"rfunc: invalid number of bind arguments (got=%d, want=%d)",
+			got, want,
+		)
+	}
+	usr.v1 = args[0].(*int32)
+	usr.v2 = args[1].(*float32)
+	usr.v3 = args[2].(*string)
+	return nil
+}
+
+func (usr *UsrStr) Func() interface{} {
+	return func() string {
+		return usr.fct(*usr.v1, *usr.v2, *usr.v3)
+	}
+}
+
+func ExampleReader_withFormulaFromUser() {
+	f, err := groot.Open("../testdata/simple.root")
+	if err != nil {
+		log.Fatalf("could not open ROOT file: %+v", err)
+	}
+	defer f.Close()
+
+	o, err := f.Get("tree")
+	if err != nil {
+		log.Fatalf("could not retrieve ROOT tree: %+v", err)
+	}
+	t := o.(rtree.Tree)
+
+	var (
+		data struct {
+			V1 int32   `groot:"one"`
+			V2 float32 `groot:"two"`
+			V3 string  `groot:"three"`
+		}
+		rvars = rtree.ReadVarsFromStruct(&data)
+	)
+
+	r, err := rtree.NewReader(t, rvars)
+	if err != nil {
+		log.Fatalf("could not create tree reader: %+v", err)
+	}
+	defer r.Close()
+
+	f64, err := r.Formula(&UsrF64{
+		rvars: []string{"one", "two", "three"},
+		fct: func(v1 int32, v2 float32, v3 string) float64 {
+			return float64(v2*10) + float64(1000*v1) + float64(100*len(v3))
+		},
+	})
+	if err != nil {
+		log.Fatalf("could not create formula: %+v", err)
+	}
+
+	fstr, err := r.Formula(&UsrStr{
+		rvars: []string{"one", "two", "three"},
+		fct: func(v1 int32, v2 float32, v3 string) string {
+			return fmt.Sprintf(
+				"%q: %v, %q: %v, %q: %v",
+				"one", v1, "two", v2, "three", v3,
+			)
+		},
+	})
+	if err != nil {
+		log.Fatalf("could not create formula: %+v", err)
+	}
+
+	f1 := f64.Func().(func() float64)
+	f2 := fstr.Func().(func() string)
+
+	err = r.Read(func(ctx rtree.RCtx) error {
+		v64 := f1()
+		str := f2()
+		fmt.Printf("evt[%d]: %v, %v, %v -> %g | %s\n", ctx.Entry, data.V1, data.V2, data.V3, v64, str)
+		return nil
+	})
+	if err != nil {
+		log.Fatalf("could not process tree: %+v", err)
+	}
+
+	// Output:
+	// evt[0]: 1, 1.1, uno -> 1311 | "one": 1, "two": 1.1, "three": uno
+	// evt[1]: 2, 2.2, dos -> 2322 | "one": 2, "two": 2.2, "three": dos
+	// evt[2]: 3, 3.3, tres -> 3433 | "one": 3, "two": 3.3, "three": tres
+	// evt[3]: 4, 4.4, quatro -> 4644 | "one": 4, "two": 4.4, "three": quatro
 }