… groot/rtree/rfunc/rfunc_bool_gen.go
… groot/rtree/rfunc/rfunc_bool_gen_test.go
… groot/rtree/rfunc/rfunc_float32_gen.go
… groot/rtree/rfunc/rfunc_float32_gen_test.go
… groot/rtree/rfunc/rfunc_float64_gen.go
… groot/rtree/rfunc/rfunc_float64_gen_test.go
… groot/rtree/rfunc/rfunc_int32_gen.go
… groot/rtree/rfunc/rfunc_int32_gen_test.go
… groot/rtree/rfunc/rfunc_int64_gen.go
… groot/rtree/rfunc/rfunc_int64_gen_test.go
… groot/rtree/rfunc/rfunc_uint32_gen.go
… groot/rtree/rfunc/rfunc_uint32_gen_test.go
… groot/rtree/rfunc/rfunc_uint64_gen.go
… groot/rtree/rfunc/rfunc_uint64_gen_test.go