introducing the rfunc.Formula protocol brings this kind of performance
improvements:

```
$> benchstat ref.txt new-rfunc.txt
name                   old time/op    new time/op    delta
FormulaFunc/f0/Func-4     238ns ± 3%       4ns ± 1%   -98.20%  (p=0.000 n=20+16)
FormulaFunc/f1/Func-4     313ns ± 2%       5ns ± 5%   -98.39%  (p=0.000 n=20+19)
FormulaFunc/f2/Func-4     313ns ± 4%       5ns ± 3%   -98.31%  (p=0.000 n=20+20)
FormulaFunc/f3/Func-4     313ns ± 2%       5ns ± 4%   -98.32%  (p=0.000 n=18+20)

name                   old alloc/op   new alloc/op   delta
FormulaFunc/f0/Func-4     40.0B ± 0%      0.0B       -100.00%  (p=0.000 n=20+20)
FormulaFunc/f1/Func-4     48.0B ± 0%      0.0B       -100.00%  (p=0.000 n=20+20)
FormulaFunc/f2/Func-4     48.0B ± 0%      0.0B       -100.00%  (p=0.000 n=20+20)
FormulaFunc/f3/Func-4     48.0B ± 0%      0.0B       -100.00%  (p=0.000 n=20+20)

name                   old allocs/op  new allocs/op  delta
FormulaFunc/f0/Func-4      2.00 ± 0%      0.00       -100.00%  (p=0.000 n=20+20)
FormulaFunc/f1/Func-4      2.00 ± 0%      0.00       -100.00%  (p=0.000 n=20+20)
FormulaFunc/f2/Func-4      2.00 ± 0%      0.00       -100.00%  (p=0.000 n=20+20)
FormulaFunc/f3/Func-4      2.00 ± 0%      0.00       -100.00%  (p=0.000 n=20+20)
```

Fixes go-hep/hep#719.